refatoração de packages/arquitetura
