+++ src/main/java/com/mercadolivre/projetointegrador/inboundorder/dto/BatchResponseDto.java
@@ -1,0 +1,4 @@
+package com.mercadolivre.projetointegrador.inboundorder.dto;
+
+public class BatchResponseDto {
+}
