+++ src/main/java/com/mercadolivre/projetointegrador/inboundorder/dto/InboundOrderRequestDto.java
@@ -1,0 +1,18 @@
+package com.mercadolivre.projetointegrador.inboundorder.dto;
+
+import com.mercadolivre.projetointegrador.inboundorder.model.InboundOrder;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class InboundOrderRequestDto {
+
+    private InboundOrder inboundOrder;
+
+}
