+++ src/main/java/com/mercadolivre/projetointegrador/inboundorder/repository/InboundOrderRepository.java
@@ -1,0 +1,11 @@
+package com.mercadolivre.projetointegrador.inboundorder.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.mercadolivre.projetointegrador.inboundorder.model.InboundOrder;
+
+@Repository
+public interface InboundOrderRepository extends JpaRepository<InboundOrder, Long> {
+
+}
