--- src/main/java/com/mercadolivre/projetointegrador/order/controller/PurchaseOrderController.java
@@ -1,8 +1,0 @@
-package com.mercadolivre.projetointegrador.order.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class PurchaseOrderController {
-
-}
