--- src/main/java/com/mercadolivre/projetointegrador/order/repository/PurchaseOrderRepository.java
@@ -1,11 +1,0 @@
-package com.mercadolivre.projetointegrador.order.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.mercadolivre.projetointegrador.order.model.PurchaseOrder;
-
-@Repository
-public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Long >{
-
-}
