--- src/main/java/com/mercadolivre/projetointegrador/order/service/PurchaseOrderService.java
@@ -1,8 +1,0 @@
-package com.mercadolivre.projetointegrador.order.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PurchaseOrderService {
-
-}
