+++ src/main/java/com/mercadolivre/projetointegrador/purchaseorder/controller/PurchaseOrderController.java
@@ -1,0 +1,8 @@
+package com.mercadolivre.projetointegrador.purchaseorder.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PurchaseOrderController {
+
+}
