+++ src/main/java/com/mercadolivre/projetointegrador/purchaseorder/repository/PurchaseOrderRepository.java
@@ -1,0 +1,11 @@
+package com.mercadolivre.projetointegrador.purchaseorder.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.mercadolivre.projetointegrador.purchaseorder.model.PurchaseOrder;
+
+@Repository
+public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Long >{
+
+}
