+++ src/main/java/com/mercadolivre/projetointegrador/purchaseorder/service/PurchaseOrderService.java
@@ -1,0 +1,8 @@
+package com.mercadolivre.projetointegrador.purchaseorder.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PurchaseOrderService {
+
+}
