+++ src/main/java/com/mercadolivre/projetointegrador/section/model/Section.java
@@ -1,0 +1,39 @@
+package com.mercadolivre.projetointegrador.section.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.mercadolivre.projetointegrador.enums.ProductType;
+import com.mercadolivre.projetointegrador.warehouse.model.Warehouse;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+@Entity
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Section implements Serializable {
+
+    private static final long serialVersionUID = 4446300291155815773L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String sectionCode;
+
+    @Enumerated(EnumType.STRING)
+    private ProductType sectionType;
+
+    @ManyToOne
+    @JsonIgnoreProperties("sections")
+    private Warehouse warehouseCode;
+
+
+    private Integer maxCapacity;
+
+}
