+++ src/main/java/com/mercadolivre/projetointegrador/section/repository/SectionRepository.java
@@ -1,0 +1,11 @@
+package com.mercadolivre.projetointegrador.section.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.mercadolivre.projetointegrador.section.model.Section;
+
+@Repository
+public interface SectionRepository extends JpaRepository<Section, Long>{
+
+}
