--- src/main/java/com/mercadolivre/projetointegrador/stock/controller/BatchController.java
@@ -1,38 +1,0 @@
-package com.mercadolivre.projetointegrador.stock.controller;
-
-import com.mercadolivre.projetointegrador.stock.model.Batch;
-import com.mercadolivre.projetointegrador.stock.service.BatchService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/batch")
-public class BatchController {
-
-    @Autowired
-    BatchService service;
-
-    @PostMapping()
-    public ResponseEntity<?> create(@RequestBody Batch batch){
-        Batch result = service.createBatch(batch);
-        return new ResponseEntity<>(result, HttpStatus.CREATED);
-    }
-
-    @GetMapping
-    public ResponseEntity<?> findAll(){
-        List<Batch> result = service.findAllBatch();
-        return new ResponseEntity<>(result, HttpStatus.OK);
-    }
-
-    @PutMapping()
-    public ResponseEntity<?> update(@RequestBody Batch batch){
-        Batch result = service.updateBatch(batch);
-        return new ResponseEntity<>(result, HttpStatus.CREATED);
-    }
-
-
-}
