--- src/main/java/com/mercadolivre/projetointegrador/stock/dto/InboundOrderRequestDto.java
@@ -1,17 +1,0 @@
-package com.mercadolivre.projetointegrador.stock.dto;
-
-import com.mercadolivre.projetointegrador.stock.model.InboundOrder;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class InboundOrderRequestDto {
-
-    private InboundOrder inboundOrder;
-
-}
