--- src/main/java/com/mercadolivre/projetointegrador/stock/dto/SectionResponseDto.java
@@ -1,4 +1,0 @@
-package com.mercadolivre.projetointegrador.stock.dto;
-
-public class SectionResponseDto {
-}
