--- src/main/java/com/mercadolivre/projetointegrador/stock/enums/ProductType.java
@@ -1,15 +1,0 @@
-package com.mercadolivre.projetointegrador.stock.enums;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum ProductType {
-    FRESH("FR", "Fresh"),
-    REFRIGERATED("RF", "Refrigerated"),
-    FROZEN("FZ", "Frozen");
-
-    private final String code;
-    private final String description;
-}
