--- src/main/java/com/mercadolivre/projetointegrador/stock/model/Warehouse.java
@@ -1,31 +1,0 @@
-package com.mercadolivre.projetointegrador.stock.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.util.List;
-
-@Entity
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Warehouse implements Serializable {
-
-    private static final long serialVersionUID = 7463206541614851461L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String code;
-
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "warehouseCode")
-    @JsonIgnoreProperties("warehouseCode")
-    private List<Section> sections;
-
-}
