+++ src/main/java/com/mercadolivre/projetointegrador/warehouse/repository/WarehouseRepository.java
@@ -1,0 +1,12 @@
+package com.mercadolivre.projetointegrador.warehouse.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.mercadolivre.projetointegrador.warehouse.model.Warehouse;
+
+
+@Repository
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long>{
+
+}
